RHBZ1187588: don't exit() from a failing pmGetConfig
An eagle-eyed mvollmer noted the ominous warnings & code for
pmGetConfig(), wherein a missing pcp.conf file was deemed a "FATAL PCP
ERROR" and resulted in a process exit().  This sort of error handling
is not appropriate within a library, as the application is better
equipped to judge the severity of the problem.  The caller must
already handle return values of empty strings, so now a missing
/etc/pcp.conf is treated just as a misformatted one, and results in
the same "" result.

QA 448 is updated to accept a rc=0 outcome from a non-crashing
src/getconfig, and the pmgetconfig.3 man page scares less.